refactor(client): improve semantic extract tabs from musicplayer fullscreen into several files
